+++ myproject/myCloudinary/cloudinaryConfig.py
@@ -1,0 +1,9 @@
+import cloudinary
+
+def cloudinaryConfig():
+    cloudinary.config( 
+    cloud_name = "odinsully", 
+    api_key = "152393343981388", 
+    api_secret = "ZJWfM-yTEFEYYCd13epWWj6tINs",
+    secure = True
+    )
